--- src/Cart/lib/prisma.ts
@@ -1,5 +1,0 @@
-import { PrismaClient } from '@prisma/client';
-
-const prisma = new PrismaClient();
-
-export default prisma;
