+++ src/routes/cart.router.ts
@@ -1,0 +1,21 @@
+// src/cart/cart.routes.ts
+import { Router } from 'express';
+import * as cartController from '../controllers/cart.controller.js';
+import { authMiddleware } from '../middlewares/auth.middleware.js';
+import { validateMiddleware } from '../middlewares/validate.middleware.js';
+import { patchCartDto } from '../dtos/cart.dto.js';
+
+export const cartRouter = Router();
+
+cartRouter.post('/', authMiddleware, cartController.createCart);
+
+cartRouter.get('/', authMiddleware, cartController.getCart);
+
+cartRouter.patch(
+  '/',
+  authMiddleware,
+  validateMiddleware({ body: patchCartDto }),
+  cartController.patchCart,
+);
+
+cartRouter.delete('/:cartItemId', authMiddleware, cartController.deleteCartItem);
