+++ src/services/cart.service.ts
@@ -1,6 +1,6 @@
 // cart.service.ts
 //import prisma from '../lib/prisma.js';
-import * as cartRepository from '../repository/cart.repository.js';
+import * as cartRepository from '../repositories/cart.repository.js';
 import type { PatchCartInput } from '../dtos/cart.dto.js';
 
 export async function createCart(userId: string) {
